[pug-lang] removed TypeInferCombinator.{c|h} and integrated into TypeInfer.{c|h}
